박종원 : 버텍스쉐이더 내용물 DX11 로 대체

+++ GameEngine/GameEngineRenderTarget.cpp
@@ -49,7 +49,7 @@ void GameEngineRenderTarget::Setting(int _Index)
 
 	if (-1 == _Index)
 	{
-		GameEngineDirectXDevice::GetContext()->OMSetRenderTargets(RenderTargetViews_.size(), &RenderTargetViews_[0], nullptr);
+		GameEngineDirectXDevice::GetContext()->OMSetRenderTargets(static_cast<UINT>(RenderTargetViews_.size()), &RenderTargetViews_[0], nullptr);
 	}
 	else
 	{
